# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage films and register users.\n[Read me](https://github.com/migalpha/kentech-films)",
        "title": "Kentech-Films",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/favourites/{film_id}": {
            "delete": {
                "description": "Allow to remove a film from favourites list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favourites"
                ],
                "summary": "Remove a film from favourites",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Remove favourites",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/films/{film_id}": {
            "get": {
                "description": "Given a film_id returns this film.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Get a specific film.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Film id",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns film",
                        "schema": {
                            "$ref": "#/definitions/film.Film"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Allow to remove a film from records.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Destroy a film from records.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Destroy film",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "error 403",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Allow to update one or many fields of films.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Allow to update one or many fields of films.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "film id",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data to update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateFilmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "error 403",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "film.Film": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "integer"
                },
                "director": {
                    "type": "string",
                    "example": "Quentin Tarantino"
                },
                "genre": {
                    "type": "string",
                    "example": "action"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "release_year": {
                    "type": "integer",
                    "example": 2009
                },
                "starring": {
                    "type": "string",
                    "example": "Brad Pitt, Christoph Waltz, Michael Fassbender"
                },
                "sypnosis": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Inglourious Basterds"
                }
            }
        },
...
        "http.emptyResponse": {
            "type": "object"
        },
        "http.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "http.updateFilmRequest": {
            "type": "object",
            "properties": {
                "director": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Quentin Tarantino"
                },
                "genre": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "action, comedy, war"
                },
                "release_year": {
                    "type": "integer",
                    "example": 2009
                },
                "starring": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Brad Pitt, Christoph Waltz, Michael Fassbender"
                },
                "sypnosis": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Inglourious Basterds"
                }
            }
        }
    }
}